refactor(command): Remove concurrency from circuit breakers
In the original crius libraries, circuit breakers carry thread pools
onto which they schedule the execution of functions.

In our use cases we would prefer the ability to control concurrent
execution ourselves.

This commit removes all usage of channels and threads from the command
module and makes circuit breaker execution completely synchronous. It
is up to the caller to determine if - and how - concurrency is
required.