fix(window): Don't assume windows are contiguous
The "driver" of window & bucket creation is currently access of the
circuit breakers, which may not guarantee contiguous windows in case
of infrequent access.

This change lets new buckets be created with a `now()` instant instead
of the assuming they are right at the cutoff threshold of the previous
bucket.

It also changes the logic for retrieving the points to use in the
current breaker state calculation to account for this by checking all
present buckets against a calculated cutoff threshold.

This change causes some incompatibilities with the rest of the code
due to changed type requirements (signed integers were previously used
in places were unsigned integers are required), this will be cleaned
up in the next commit.